+++ CRM/Mautic/Exception/APIException.php
@@ -1,9 +1,7 @@
 <?php
 /**
  * @file
- * 
+ *
  */
 
-abstract class CRM_Mautic_Exception_APIException extends Exception {
-
-}
+class CRM_Mautic_Exception_APIException extends Exception {}