+++ sql/mauticwebhook_install.sql
@@ -11,7 +11,7 @@
 --
 -- /*******************************************************
 -- *
--- * Clean up the existing tables
+-- * Clean up the existing tables - this section generated from drop.tpl
 -- *
 -- *******************************************************/
 
@@ -39,7 +39,7 @@ CREATE TABLE `civicrm_mauticwebhook` (
   `webhook_trigger_type` varchar(255) COMMENT 'Mautic Webhook trigger event type.',
   `activity_id` int unsigned COMMENT 'FK to Contact',
   `contact_id` int unsigned COMMENT 'FK to Contact',
-  `created_date` timestamp NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'When the webhook was first received by CiviCRM',
+  `created_date` timestamp DEFAULT CURRENT_TIMESTAMP COMMENT 'When the webhook was first received by CiviCRM',
   `processed_date` timestamp NULL DEFAULT NULL COMMENT 'Has this webhook been processed in CiviCRM',
   PRIMARY KEY (`id`)
-);
+) ENGINE=InnoDB;